+++ IEX_29id/devices/undulator.py
@@ -1,5 +1,4 @@
-from epics import caget, caput
-from time import sleep
+from IEX_29id.devices.energy import Switch_IDMode
 
 
 
@@ -16,3 +15,9 @@ def ID_State2Mode(which,mode):
 #replace with    ?
 #   ID_Mode_list=['RCP','LCP','V','H','HN']
 #   ID_Mode=ID_Mode_list[caget("ID29:ActualMode")]
+
+def polarization(which):
+    """
+    Change beam polarization: which = \"H\", \"V\", \"RCP\" or \"LCP\"
+    """
+    Switch_IDMode(which)
